+++ addon/components/variable-number/number.ts
@@ -17,7 +17,7 @@ import {
   MINIMUM_VALUE_PNODE_KEY,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/utils/constants';
 import { isBlank } from '@ember/utils';
-import n2words from 'n2words';
+import { numberToWords } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/utils/number-to-words';
 
 type Args = {
   getPos: () => number | undefined;
@@ -35,6 +35,7 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
   @tracked errorMessage = '';
   @service declare intl: intlService;
   cursorPositionKeyDown: number | null = null;
+  @tracked infoTooltipOpen = false;
 
   focus(element: HTMLInputElement) {
     element.focus();
@@ -52,7 +53,7 @@ export default class VariableNumberPluginNumberComponent extends Component<Args>
     if (!this.writtenNumber) {
       return value;
     } else {
-      return n2words(Number(value), { lang: 'nl' });
+      return numberToWords(Number(value), { lang: 'nl' });
     }
   }