+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Addition of documentation on the article-structure plugin
+
+### Fixed
+- When a number is too big to be converted to words, display it numerically.
 ### Dependencies
 - Bumps `@typescript-eslint/parser` from 5.60.0 to 5.61.0
 - Bumps `@tsconfig/ember` from 1.0.1 to 3.0.0